Auto-generate the YAML file that describes the version (#11)
Currently we have separate, almost entirely identical files,
`snapcraft-lts.yaml` and `snapcraft-stable.yaml`, and on CI we copy the
appropriate one to `snapcraft.yaml`. This requires keeping the patch
versions of both LTS and stable up to date, as both are hard-coded in
their respective files.

Instead, we can automatically generate `snapcraft.yaml` based on the
release type (`lts` or `stable`) and `versions.json`. This allows us to
select the most recent LTS patch or stable version without needing to
edit anything directly.

Co-authored-by: Pablo Zubieta <8410335+pabloferz@users.noreply.github.com>